Restructure functional test directory
