+++ tests/functional/forklift/test_legacy.py
@@ -10,6 +10,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+import pytest
+
 
 def test_incorrect_post_redirect(webtest):
     """
@@ -27,3 +29,20 @@ def test_incorrect_post_redirect(webtest):
 
     assert "location" in resp.headers
     assert resp.headers["location"] == "http://localhost/legacy/"
+
+
+@pytest.mark.parametrize("action", ["submit", "submit_pkg_info"])
+def test_removed_upload_apis(webtest, action):
+    resp = webtest.post(f"/legacy/?:action={action}", status=410)
+    assert resp.status == (
+        "410 Project pre-registration is no longer required or supported, "
+        "upload your files instead."
+    )
+
+
+def test_remove_doc_upload(webtest):
+    resp = webtest.post("/legacy/?:action=doc_upload", status=410)
+    assert resp.status == (
+        "410 Uploading documentation is no longer supported, we recommend "
+        "using https://readthedocs.org/."
+    )
