+++ tests/functional/legacy/api/test_pypi.py
@@ -10,25 +10,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import pytest
-
-
-@pytest.mark.parametrize("action", ["submit", "submit_pkg_info"])
-def test_removed_upload_apis(webtest, action):
-    resp = webtest.post(f"/legacy/?:action={action}", status=410)
-    assert resp.status == (
-        "410 Project pre-registration is no longer required or supported, "
-        "upload your files instead."
-    )
-
-
-def test_remove_doc_upload(webtest):
-    resp = webtest.post("/legacy/?:action=doc_upload", status=410)
-    assert resp.status == (
-        "410 Uploading documentation is no longer supported, we recommend "
-        "using https://readthedocs.org/."
-    )
-
 
 def test_doap(webtest):
     resp = webtest.get("/pypi?:action=doap&name=foo&version=1.0", status=410)
